Move PDataTFStream back inside DicomService.cs so it can be reviewed properly
